Move DocumentConversionServiceImpl to Core
